Crate rusoto_shield

AWS Shield Advanced
This is the AWS Shield Advanced API Reference. This guide is for developers who need detailed information about the AWS Shield Advanced API actions, data types, and errors. For detailed information about AWS WAF and AWS Shield Advanced features and an overview of how to use the AWS WAF and AWS Shield Advanced APIs, see the AWS WAF and AWS Shield Developer Guide.

If you're using the service, you're probably looking for ShieldClient and Shield.

Structs

AssociateDRTLogBucketRequest

AssociateDRTLogBucketResponse

AssociateDRTRoleRequest

AssociateDRTRoleResponse

AttackDetail
The details of a DDoS attack.

AttackProperty
Details of the described attack.

AttackSummary
Summarizes all DDoS attacks for a specified time period.

AttackVectorDescription
Describes the attack.

Contributor
A contributor to the attack and their contribution.

CreateProtectionRequest

CreateProtectionResponse

CreateSubscriptionRequest

CreateSubscriptionResponse

DeleteProtectionRequest

DeleteProtectionResponse

DeleteSubscriptionRequest

DeleteSubscriptionResponse

DescribeAttackRequest

DescribeAttackResponse

DescribeDRTAccessRequest

DescribeDRTAccessResponse

DescribeEmergencyContactSettingsRequest

DescribeEmergencyContactSettingsResponse

DescribeProtectionRequest

DescribeProtectionResponse

DescribeSubscriptionRequest

DescribeSubscriptionResponse

DisassociateDRTLogBucketRequest

DisassociateDRTLogBucketResponse

DisassociateDRTRoleRequest

DisassociateDRTRoleResponse

EmergencyContact
Contact information that the DRT can use to contact you during a suspected attack.

GetSubscriptionStateRequest

GetSubscriptionStateResponse

Limit
Specifies how many protections of a given type you can create.

ListAttacksRequest

ListAttacksResponse

ListProtectionsRequest

ListProtectionsResponse

Mitigation
The mitigation applied to a DDoS attack.

Protection
An object that represents a resource that is under DDoS protection.

ShieldClient
A client for the AWS Shield API.

SubResourceSummary
The attack information for the specified SubResource.

Subscription
Information about the AWS Shield Advanced subscription for an account.

SummarizedAttackVector
A summary of information about the attack.

SummarizedCounter
The counter that describes a DDoS attack.

TimeRange
The time range.

UpdateEmergencyContactSettingsRequest

UpdateEmergencyContactSettingsResponse

UpdateSubscriptionRequest

UpdateSubscriptionResponse

Enums

AssociateDRTLogBucketError
Errors returned by AssociateDRTLogBucket

AssociateDRTRoleError
Errors returned by AssociateDRTRole

CreateProtectionError
Errors returned by CreateProtection

CreateSubscriptionError
Errors returned by CreateSubscription

DeleteProtectionError
Errors returned by DeleteProtection

DeleteSubscriptionError
Errors returned by DeleteSubscription

DescribeAttackError
Errors returned by DescribeAttack

DescribeDRTAccessError
Errors returned by DescribeDRTAccess

DescribeEmergencyContactSettingsError
Errors returned by DescribeEmergencyContactSettings

DescribeProtectionError
Errors returned by DescribeProtection

DescribeSubscriptionError
Errors returned by DescribeSubscription

DisassociateDRTLogBucketError
Errors returned by DisassociateDRTLogBucket

DisassociateDRTRoleError
Errors returned by DisassociateDRTRole

GetSubscriptionStateError
Errors returned by GetSubscriptionState

ListAttacksError
Errors returned by ListAttacks

ListProtectionsError
Errors returned by ListProtections

UpdateEmergencyContactSettingsError
Errors returned by UpdateEmergencyContactSettings

UpdateSubscriptionError
Errors returned by UpdateSubscription

Traits

Shield
Trait representing the capabilities of the AWS Shield API. AWS Shield clients implement this trait.