Struct rusoto_shield::ShieldClient

pub struct ShieldClient { /* fields omitted */ }

A client for the AWS Shield API.

Methods

impl ShieldClient

pub fn new(region: Region) -> ShieldClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ShieldClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,