Trait rusoto_shield::Shield

pub trait Shield {
    fn create_protection(
        &self, 
        input: CreateProtectionRequest
    ) -> RusotoFuture<CreateProtectionResponse, CreateProtectionError>;

    fn create_subscription(
        &self
    ) -> RusotoFuture<CreateSubscriptionResponse, CreateSubscriptionError>;

    fn delete_protection(
        &self, 
        input: DeleteProtectionRequest
    ) -> RusotoFuture<DeleteProtectionResponse, DeleteProtectionError>;

    fn delete_subscription(
        &self
    ) -> RusotoFuture<DeleteSubscriptionResponse, DeleteSubscriptionError>;

    fn describe_attack(
        &self, 
        input: DescribeAttackRequest
    ) -> RusotoFuture<DescribeAttackResponse, DescribeAttackError>;

    fn describe_protection(
        &self, 
        input: DescribeProtectionRequest
    ) -> RusotoFuture<DescribeProtectionResponse, DescribeProtectionError>;

    fn describe_subscription(
        &self
    ) -> RusotoFuture<DescribeSubscriptionResponse, DescribeSubscriptionError>;

    fn get_subscription_state(
        &self
    ) -> RusotoFuture<GetSubscriptionStateResponse, GetSubscriptionStateError>;

    fn list_attacks(
        &self, 
        input: ListAttacksRequest
    ) -> RusotoFuture<ListAttacksResponse, ListAttacksError>;

    fn list_protections(
        &self, 
        input: ListProtectionsRequest
    ) -> RusotoFuture<ListProtectionsResponse, ListProtectionsError>;
}

Trait representing the capabilities of the AWS Shield API. AWS Shield clients implement this trait.

Required Methods

fn create_protection(
&self,
input: CreateProtectionRequest
) -> RusotoFuture<CreateProtectionResponse, CreateProtectionError>

Enables AWS Shield Advanced for a specific AWS resource. The resource can be an Amazon CloudFront distribution, Elastic Load Balancing load balancer, Elastic IP Address, or an Amazon Route 53 hosted zone.

fn create_subscription(
&self
) -> RusotoFuture<CreateSubscriptionResponse, CreateSubscriptionError>

Activates AWS Shield Advanced for an account.

fn delete_protection(
&self,
input: DeleteProtectionRequest
) -> RusotoFuture<DeleteProtectionResponse, DeleteProtectionError>

Deletes an AWS Shield Advanced Protection.

fn delete_subscription(
&self
) -> RusotoFuture<DeleteSubscriptionResponse, DeleteSubscriptionError>

Removes AWS Shield Advanced from an account. AWS Shield Advanced requires a 1-year subscription commitment. You cannot delete a subscription prior to the completion of that commitment.

fn describe_attack(
&self,
input: DescribeAttackRequest
) -> RusotoFuture<DescribeAttackResponse, DescribeAttackError>

Describes the details of a DDoS attack.

fn describe_protection(
&self,
input: DescribeProtectionRequest
) -> RusotoFuture<DescribeProtectionResponse, DescribeProtectionError>

Lists the details of a Protection object.

fn describe_subscription(
&self
) -> RusotoFuture<DescribeSubscriptionResponse, DescribeSubscriptionError>

Provides details about the AWS Shield Advanced subscription for an account.

fn get_subscription_state(
&self
) -> RusotoFuture<GetSubscriptionStateResponse, GetSubscriptionStateError>

Returns the SubscriptionState, either Active or Inactive.

fn list_attacks(
&self,
input: ListAttacksRequest
) -> RusotoFuture<ListAttacksResponse, ListAttacksError>

Returns all ongoing DDoS attacks or all DDoS attacks during a specified time period.

fn list_protections(
&self,
input: ListProtectionsRequest
) -> RusotoFuture<ListProtectionsResponse, ListProtectionsError>

Lists all Protection objects for the account.